// src/interfaces/UsuarioInterface.tsx
// Generated by https://quicktype.io

export interface LoginData {
    email: string;
    password: string;
}

export interface RegisterData {
    name: string;
    email: string;
    password: string;
}

export interface LoginResponse {
    usuario: Usuario;
    token:   string;
}

export interface CoordinatesData {
    uid: string;
    lat:   number;
    lng:   number;
}

export interface Usuario {
    nombre: string;
    correo: string;
    rol:    string;
    estado: boolean;
    google: boolean;
    latitud: number;
    longitud: number;
    uid:    string;
}